+++ spec/classes/characteristics.test.js
@@ -3,7 +3,15 @@ import { gg, factory } from '../helpers'
 let hero = factory.hero()
 
 describe('characteristics', () => {
-  test('should initialize with valid characteristics', () => {
+  test('should initialize without characteristics instances', () => {
+    let characteristics = new gg.class.Characteristics({
+      characteristics: [
+        {name: 'MyCharacteristic', value: 1}
+      ]
+    })
+    expect(characteristics.getCharacteristics().length).toBe(3) // MyCharacteristic characteristic and life & defense defaults
+  })
+  test('should initialize with characteristics instances', () => {
     let characteristics = new gg.class.Characteristics({
       characteristics: [
         new gg.class.Characteristic({name: 'MyCharacteristic', value: 1})
@@ -11,8 +19,16 @@ describe('characteristics', () => {
     })
     expect(characteristics.getCharacteristics().length).toBe(3) // MyCharacteristic characteristic and life & defense defaults
   })
-  test('should\'nt initialize with invalid characteristics', () => {
-    expect(() => new gg.class.Characteristics({characteristics: [{name: 'Dummy characteristic', value: 1}]})).toThrow(new Error('Invalid characteristics'))
+  test('should set characteristics', () => {
+    let characteristics = new gg.class.Characteristics({
+      characteristics: [
+        new gg.class.Characteristic({name: 'MyCharacteristic', value: 1})
+      ]
+    })
+    characteristics.setCharacteristics([
+      {name: 'MyCharacteristic', value: 2}
+    ])
+    expect(characteristics.getCharacteristics().length).toBe(3) // MyCharacteristic characteristic and life & defense defaults
   })
   test('should increase a characteristic', () => {
     hero.characteristics.increase('defense', 1)