+++ spec/classes/vest.test.js
@@ -1,4 +1,5 @@
 import { factory } from '../helpers'
+import { characteristic as characteristicTypes } from 'constants'
 
 let helmet = factory.helmet()
 
@@ -8,6 +9,17 @@ describe('vest', () => {
       factory.effect({
         characteristic: factory.health()
       })
+    ])
+    expect(helmet.getEffects().length).toBe(1)
+  })
+  test('should be able to set effects', () => {
+    helmet.setEffects([
+      {
+        characteristic: {
+          name: characteristicTypes.DEFENSE,
+          value: 10
+        }
+      }
     ])
     expect(helmet.getEffects().length).toBe(1)
   })