+++ src/classes/index.js
@@ -1,5 +1,5 @@
-export * from './items'
 export * from './general'
+export * from './items'
 export * from './character'
 export * from './shop'
 export * from './quest'