+++ src/classes/items/vest/Vest.js
@@ -1,15 +1,19 @@
-import { Item } from '../../general'
+import { Item, Characteristic } from 'classes'
+import { parseToInstance } from 'helpers'
 
 class Vest extends Item {
   constructor ({name, type, slotType, effects}) {
     super({name, type, slotType})
-    this._effects = effects
+    this._effects = this._parseEffects(effects)
   }
   setEffects (effects) {
-    this._effects = effects
+    this._effects = this._parseEffects(effects)
   }
   getEffects () {
     return this._effects
+  }
+  _parseEffects (effects) {
+    return parseToInstance(Characteristic, effects)
   }
 }