+++ src/classes/items/weapon/Weapon.js
@@ -1,10 +1,10 @@
-import { Item } from '../../general'
+import { Item, Attack } from 'classes'
 import { clone } from 'helpers'
 
 class Weapon extends Item {
   constructor ({name, type, weight, slotType, attacks, floorDamage = 0}) {
     super({name, type, weight, slotType})
-    this._attacks = attacks
+    this._attacks = this._parseAttacks(attacks)
     this._floorDamage = floorDamage
   }
   getAttacks () {
@@ -15,13 +15,16 @@ class Weapon extends Item {
     })
   }
   setAttacks (attacks) {
-    this._attacks = attacks
+    this._attacks = this._parseAttacks(attacks)
   }
   setFloorDamage (floorDamage) {
     this._floorDamage = floorDamage
   }
   getFloorDamage () {
     return this._floorDamage
+  }
+  _parseAttacks (attacks) {
+    return (attacks && attacks.map(attack => attack instanceof Attack ? attack : new Attack(attack))) || []
   }
 }