+++ src/helpers/index.js
@@ -1,1 +1,2 @@
 export clone from './clone'
+export parseToInstance from './parseToInstance'