+++ src/helpers/parseToInstance.js
@@ -1,0 +1,3 @@
+const parseToInstance = (Class, list) => list.map(item => item instanceof Class ? item : new Class(item))
+
+export default parseToInstance